+++ .idea/ICESP4.iml
@@ -7,8 +7,9 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="sqlite-jdbc-3.44.1.0" level="project" />
+    <orderEntry type="library" exported="" name="sqlite-jdbc-3.44.1.0" level="project" />
     <orderEntry type="library" name="slf4j-simple-2.0.9" level="project" />
     <orderEntry type="library" name="slf4j-api-2.0.9" level="project" />
+    <orderEntry type="library" exported="" name="xerial.sqlite.jdbc" level="project" />
   </component>
 </module>
