+++ .idea/libraries/xerial_sqlite_jdbc.xml
@@ -1,0 +1,12 @@
+<component name="libraryTable">
+  <library name="xerial.sqlite.jdbc" type="repository">
+    <properties maven-id="org.xerial:sqlite-jdbc:3.47.0.0" />
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/lib/sqlite-jdbc-3.47.0.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$PROJECT_DIR$/lib/sqlite-jdbc-3.47.0.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES />
+  </library>
+</component>
