+++ doc-diagram/Klassediagram.puml
@@ -62,12 +62,10 @@ class Quiz {
 - List<Question> questions
 - int currentQuestionIndex
 + Quiz(Player player)
-+ loadQuestions()
-+ start()
-+ nextQuestion()
-+ submitAnswer(int answer)
-+ calculateScore() int
-+ showResults()
++ loadNextQuestion()
++ boolean checkAnswer(int userAnswer, Question question)
++ boolean isQuizComplete()
++ getPlayer()
 }
 
 class TextUI {
