+++ src/Main.java
@@ -1,10 +1,9 @@
-import util.*;
+import domain.Database;
+import util.TextUI;
 
-public class Main {
     public static void main(String[] args) {
         Database.startDatabase();
 
         TextUI ui = new TextUI();
         ui.run();
     }
-}
